feat: Exceptions handling.

+++ src/main/java/com/luckraw/kwcommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.luckraw.kwcommerce.services;
 import com.luckraw.kwcommerce.dto.ProductDTO;
 import com.luckraw.kwcommerce.entities.Product;
 import com.luckraw.kwcommerce.repositories.ProductRepository;
+import com.luckraw.kwcommerce.services.exceptions.DatabaseException;
+import com.luckraw.kwcommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -18,7 +23,8 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
 
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 
         return new ProductDTO(product);
     }
@@ -44,18 +50,33 @@ public class ProductService {
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
 
-        Product product = repository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
+        try {
+            Product product = repository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
 
-        repository.save(product);
+            repository.save(product);
 
-        return new ProductDTO(product);
+            return new ProductDTO(product);
+
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
 
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
+        try {
+            repository.deleteById(id);
+
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
